Move match & resolution logic from finder to route part handler
